Added v4 to the Dashboard Version: * results.show_date now toggles date visibility. * results.fields now contains a list of fields (strings) to display.
If validation of a dashboard fails, it still displays, but the warning is logged to the console.

Admins can now see and modify the dashboard version from the UI.

Multibox now supports a selectValue method that hides the popup on selection.  Future enhancements will allow this selection method to be externally defined.

Multibox popups now correctly appear next to the source field.

When a pivot fails, more data is written to the log about what went wrong.