Bind background element click to clear the current selection.
Currently, ESCAPE clears the current widget selection which is
not very discoverable. Also set up event bindings for background
element clicks.

This is a bit ugly due to event propagation - instead of adding
stopPropagation to all existing click events, examine the clicked-on
element to see if it's a known background element.